Completed the basic ddOutput generator and started testing it. Fixed a small bug in splitSet.cpp for cos intervals when a single value was desired.
fml file read hits three unknown lines (avv ddscat output versions), starting near finite target. Need to add these to ddOutputSingle parsers.

Need to complete + fix F->S->P calculation. It is slightly off, and P31 and P41 have the wrong sign.

Need a ddOutput extractor program.

ddUtil needs a program_options handler to set ddscat version tag names. Tagging routine in ddscat-data needs to be updated for this. 
ddscat-postprocess needs to add this also.

fixed environment variable retrieval null exception when the environ var was not found. still need to get rid of nearby compiler warning.

may need a new reader in ddOutputSingle that reads in after a character, like an '=' sign. beta, theta, phi angle reads are fragile, and the spacing may change in later ddscat versions. I fixed the beta, phi reading bug.

Still need ddOutput hashing and naming function.

Also should save original (before reprocessing from fml) data for Pnn.



git-svn-id: file:///mnt/svn/rtmath@1097 1d52797c-35fc-48e8-8e80-2db72aa187e2

+++ rtmath/rtmath/src/ddOutputGenerator.cpp
@@ -19,7 +19,7 @@
 
 namespace rtmath {
 	namespace ddscat {
-
+		/*
 		ddOutputGenerator::ddOutputGenerator() {}
 
 		ddOutputGenerator::ddOutputGenerator(boost::shared_ptr<ddOutput> source) : src(source)
@@ -98,7 +98,7 @@ namespace rtmath {
 			for (auto &sca : res->scas)
 			{
 				res->weights.insert(std::pair<boost::shared_ptr<ddOutputSingle>, float>
-					(sca, wts.getWeight(sca->beta(), sca->theta(), sca->phi()) ));
+					(sca, (float) wts.getWeight(sca->beta(), sca->theta(), sca->phi()) ));
 			}
 
 			/// \todo Construct the ddOutputSingle here
@@ -135,13 +135,13 @@ namespace rtmath {
 			for (auto &sca : res->scas)
 			{
 				res->weights.insert(std::pair<boost::shared_ptr<ddOutputSingle>, float>
-					(sca, wts.getWeight(sca->beta(), sca->theta(), sca->phi()) ));
+					(sca, (float) wts.getWeight(sca->beta(), sca->theta(), sca->phi()) ));
 			}
 
 			/// \todo Construct the ddOutputSingle here
 			throw debug::xUnimplementedFunction();
 		}
-		
+		*/
 
 		/*
 		ddOutputEnsembleGaussian::ddOutputEnsembleGaussian(double sigma, size_t coord_varying)
